Move assertion checks to runtime_failures
